refactor : 커스텀 Exception 위치 변경, CarInfo의 Name, Position 원시값으로 변경
CarInfo의 Name, Position에서 단순히 get 역할만 수행하므로 원시값 int, String으로 변경
커스텀 Exception의 의존은 domain에서만 사용되므로 domain 패키지로 이동
