+++ src/main/java/racingcar/domain/Cars.java
@@ -1,7 +1,5 @@
 package racingcar.domain;
 
-import racingcar.exceptions.InvalidCarNameException;
-import racingcar.exceptions.InvalidPositionException;
 import racingcar.utils.RandomUtils;
 
 import java.util.ArrayList;
