+++ src/test/java/racingcar/domain/NameTest.java
@@ -1,7 +1,6 @@
 package racingcar.domain;
 
 import org.junit.jupiter.api.Test;
-import racingcar.exceptions.InvalidCarNameException;
 
 import static org.assertj.core.api.Assertions.*;
